SuspiciousPhoneTransferController

+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousCardTransferRepository.java
@@ -4,6 +4,8 @@ import com.bank.antifraud.model.SuspiciousCardTransfer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface SuspiciousCardTransferRepository extends JpaRepository<SuspiciousCardTransfer, Integer> {
+public interface SuspiciousCardTransferRepository extends JpaRepository<SuspiciousCardTransfer, BigInteger> {
 }
