+++ antifraud/src/main/java/com/bank/antifraud/repository/SuspiciousPhoneTransferRepository.java
@@ -4,6 +4,8 @@ import com.bank.antifraud.model.SuspiciousPhoneTransfer;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
+import java.math.BigInteger;
+
 @Repository
-public interface SuspiciousPhoneTransferRepository extends JpaRepository<SuspiciousPhoneTransfer, Integer> {
+public interface SuspiciousPhoneTransferRepository extends JpaRepository<SuspiciousPhoneTransfer, BigInteger> {
 }
